requirements.txt
opencv-python
numpy
pytesseract
Pillow
matplotlib
scikit-learn
PyMaxflow 